[pulse] bail early instead of introducing Q.undef
Summary:
The presence of `Q.undef` has a tendency to turn into unsatisfiable
formulas, as eg `undef = 1` is false, as is `undef = undef`. We can be
more "efficient" and recognise it as a contradiction when it makes sense
as early as possible.

There's only one place where we (knowingly at least) introduce
`Q.undef`: if multiplications generate constants that are too big.
Changing this to a more conservative `Q.t option` was causing too many
cumbersome changes so I left it as-is for now.

Reviewed By: fantazio

Differential Revision: D35434366

fbshipit-source-id: a270bd7693fa5a92b71edf8a89fb45cc2a663bb5