[pulse] cleanup summaries to remove unreachable addresses from inside attributes
Summary:
Let's not keep too much garbage around. We should even clean up
addresses that are not in the *precondition* as they are the only ones
that have a chance of having more taint at call sites than they had
within the callee, but for now let's just remove addresses that are
truly dead.

Reviewed By: skcho

Differential Revision: D35496294

fbshipit-source-id: 99d0ce4067666c80d29044e43d98b9c95b528081