[sledge] Report steps to reach target of goal-directed analysis
Summary:
Include the number of steps taken to reach the goal in reporting output for
goal-directed analysis, rather than just whether or not the goal was reached.

This will allow us to detect regressions in the goal-directed exploration
algorithm, and includes a simple test to check that we take the shortest
available path to a given goal.

Reviewed By: jberdine

Differential Revision: D35614228

fbshipit-source-id: 1574d4b858d43717121915722842194cb3e38d59