[erl-frontend] Refactor specs in Erlang AST a bit
Summary:
Move function around a bit so that the `spec` type is independent of the rest of the AST. This cleans up the code a bit, and will allow us to put specs in `ProcAttributes` without having circular dependencies.

Also remove some logging because now we have tests for types/specs.

Reviewed By: mmarescotti

Differential Revision: D35358026

fbshipit-source-id: d23d510cf1f83e40ac37bdd68e2b771b98ddc333