[erl-frontend] Do not add CRCs to anon names
Summary:
Anonymous functions are given names built from the name of the
surrounding named function. Previously, the name of the surrounding
function was obtained with `Procname.to_filename`, which appends a CRC.
That CRC is not needed.

Note that replacing `Procname.to_filename` by `Procname.to_string` also
removes the module name. This is intended as it is redundant: the module
of the enclosing named function is the same as the name of the anonymous
function. As a consequence, we don't need to get rid of ':'.

Reviewed By: kren1

Differential Revision: D35746158

fbshipit-source-id: 4dc58ad7b217d65cfc9b16c89f28ea8109d62150