[unnecessary copy] Keep arbitrary value as source address of copy
Summary:
In order to check if the source object is changed or not, it keeps source values of copy constructor
when possible.  Before this diff, it kept them only when logical/program variable is given as
source, so it could not reason about the source is the other expression, e.g. field access
`this.field`.

This diff keeps arbitrary value as source address of copy, by applying `eval` to the source
expression.

Reviewed By: ezgicicek

Differential Revision: D35965767

fbshipit-source-id: 29ddee8bc3e67a1b08bb8b82202dbba6a9d6f1fa