[erl-frontend] Remove fp_ prefix from tests that are not false negative
Summary: There are two Bad test where we report an error as expected (true positive), but for the wrong reason (unsupported construct). Previously these were marked as false negative, but we discussed that these shouldn't be FN. There are corresponding tests which for the same reason are false positives and are marked so.

Reviewed By: mmarescotti

Differential Revision: D35960008

fbshipit-source-id: dba6baf24e6317ed09190ad0e907a2cce547b16c