[FCP] Export is_copy_assignment to Procname for CPPMethod
Summary: This will be useful in the next diff when we extend unnecessary copy analysis to catch copies caused by copy assignment. Notice that it doesn't suffice to just check for `operator=` for the procname because move assignment also has the same name. So we resort to getting this info from clang.

Reviewed By: skcho

Differential Revision: D36008191

fbshipit-source-id: 02966c683e1e466be91c51bf3cde8a499b9067cb