[pulse taint][c++] Import quandary/basics.cpp
Summary:
If a C++ function returns a structured value, the frontend adds an additional return parameter. Instead of tainting the return value, we want to taint the return parameter (which is the last parameter of the function).

We still have 4 FN and 1 FP left to fix in basics.cpp which we will address in the future diff. These tests use a combination of variables / their addresses and their dereferences. For example, we have some value as a source, but we pass its dereference to a sink. At the moment Pulse does not taint an address or dereference for a tainted value.

Reviewed By: dulmarod

Differential Revision: D36196478

fbshipit-source-id: 57e6ae9481ba45afbec63736209497719a126146