[erl-frontend] Parse unknown types and treat them as Any
Summary:
Previously, if there was at least one unknown type in a spec, we skipped the whole spec. However, most of the time we either need the types of the arguments and ignore the return value, or the other way around. With this diff, unsupported types are parsed as `Unsupported`, which is later treated as `Any` (approximation) so that the whole spec does not need to be skipped. I added a separate `Unsupported` so that we can distinguish it from `Any` (which is a valid spec) if needed later.

See also test case where it is useful: `lists:partition` has complex arguments (function types), but it's return value is simple.

Reviewed By: thizanne

Differential Revision: D36169178

fbshipit-source-id: 1f8d961a079d564c1193a2d3648bf06bc66658c2