[callgraph] build and output file-level call graph
Summary:
The file-level call graph is constructed by lifting the static call information on procedures in the capture DB to the level of file nodes.
In addition, the file-level graph has edges labelled by the number of distinct target-file procedures called from any procedure in the source file.

This information will be used in another diff for static scheduling over distributed analysis workers.

Reviewed By: ezgicicek

Differential Revision: D36307100

fbshipit-source-id: 73fd21bf01eae8b2746d2f893000353b1d2bc172