[sledge] Output a witness trace when an alarm or goal is reached
Summary:
Track execution histories during analysis, so as to output a witness trace by
which the alarm/goal can be reached when analysis terminates.

This diff tracks the DAG of joined execution traces that reached each point
on the frontier of the symbolic execution, but currently outputs only a
single (arbitrarily-chosen) linear trace upon termination.

Reviewed By: jberdine

Differential Revision: D36414800

fbshipit-source-id: c43f245fef28c85d5c8ad1a8c7848e6ea581c64a