[sledge] Improve trace printing output, refactor goal-update logic
Summary:
Improve the printed output of witness traces by including source locations
for block terminators and appending the last call that reaches a goal to the
history trace by which that goal was reached.  Also refactors out checking if
goals are reached (and reporting if so) out of the goal-update procedure in
the process.

Reviewed By: jberdine

Differential Revision: D36476207

fbshipit-source-id: 9d3b889ebdec76e614f7ce34ef2f313e5a9b106c