[pulse][taint][report] Print data flows from/to a given source/sink
Summary:
Add new `--pulse-report-flows-from-taint-source procname` and
`--pulse-report-flows-to-taint-sink procname` flags, to be used in
conjunction with `infer report`.
These will read in a JSON report, filter its reported issues, and print
the results to `stdout`.

The filtering is on:
  - the issue type;
  - the supplied `procname` of the source or sink.

More precisely, `--pulse-report-flows-from-taint-source procname` will
select only `SENSITIVE_DATA_FLOW`s which originate from `procname`.
Similarly, `--pulse-report-flows-to-taint-sink procname` selects
`DATA_FLOWS_TO_SINK`s which end at `procname`.
This means that the two flags should not be used together - a usage error is
displayed if the user attempts this.

This commit relies on exposing a printer for JSON issues
(`IssuesTest.pp_custom_of_report`) in the corresponding `.mli` file
so that it can parse the JSON, process the result, and re-print it without
writing to an intermediate file.

Reviewed By: skcho

Differential Revision: D36703124

fbshipit-source-id: d6fd6297556e5add595a399e7593ed4d3bb81075