[pulse][taint] Report data flows from allocations to taint sinks
Summary:
Instead of reporting a "data flow to sink" advisory alongside taint
errors, report one for every path which is discovered from an allocation to a sink.

This requires changing the `expr`/`source` fields from the `FlowToTaintSink` diagnostic.
They have been combined to a single `source` field which contains a pair: the expression that
flows to the taint sink *and* its history, derived from its allocation attribute in Pulse state.

Also add a very basic test (`DataFlowToSink.m`) which creates an object
and passes it through a few dummy functions, before passing it to a taint sink.

Reviewed By: dulmarod

Differential Revision: D36810539

fbshipit-source-id: d7dd337a8685fd80d46a45fe22de3ba7da50c59c