[erl-frontend] Simplify tests with macro
Summary: Introduce macros instead of `warn(1)` in a common file to make tests more readable and uniform. This diff applies it to tests in `pulse/features`, with the exception of list comprehensions, because that would lead to false positives. I plan to do the rest in follow-up diffs to avoid large diffs and conflicts.

Reviewed By: rgrig

Differential Revision: D36627305

fbshipit-source-id: 3e4c65340c2537107eab888a26f9acf94f234b9f