[pulse][taint] Add a trace to the `TaintSanitized` attribute
Summary:
Augment the `TaintSanitized` Pulse attribute with a trace, and feed it through appropriately when the attribute is used.
This will permit clearer reporting of data flows to sinks - in particular, we would like to notify the users if there
is a sanitizer along the flow.
Having a trace available in the Pulse state will permit clear reporting of where this sanitizer interacts with the flow.

Reviewed By: skcho

Differential Revision: D37069756

fbshipit-source-id: 1f2273a44a8a6aa395a01aa99b77e55201df43bb