[pulse][taint] Consider sanitizers when reporting data flows to sinks
Summary:
When reporting a data flow to a given sink, check if there is a sanitizer somewhere in the flow - if so, report it.
We have to be careful to inspect policies: if there is no policy which allows the sanitizer to protect data for the
given sink, we should not report it.
Also, without considering sources we cannot be sure if an intervening sanitizer sufficiently protects data - the report
makes clear that the data *might* be sanitized.
Users can `infer explore` to examine the trace of where the data passes through the sanitizer.

Reviewed By: skcho

Differential Revision: D37070634

fbshipit-source-id: 1ba8c7a6dd30a24dea7d55c6ec58f5bc5c6795cc