[sledge] Avoid inlining traced functions; refactor CLI arguments
Summary:
Add a `preserve_fns` parameter to `Frontend.translate`, which accepts some
list of functions for LLVM to avoid inlining or internalizing.

When a goal trace is provided at the command line, we now pass it to the
frontend via this parameter.  This change requires a modest refactor of the
CLI, in order to provide a shared argument (`-goal-trace`) to two
subcommands (`translate` and `analyze`).

Also adds a test `noinline.c` to confirm that traced functions are preserved
even when they are inlinable, and a corresponding error case `inline_err.c`.

Reviewed By: jberdine

Differential Revision: D37138221

fbshipit-source-id: d9ea0dc88c3afddc082c8978ab5c0adc57e08ceb