[pulse taint][c++] Import quandary/unknown_code.cpp
Summary: We did not report taint error on `skip_indirect_bad` because we were dropping `PropagateTaintFrom` attribute when constructing the summary if the `taint_in` addresses were not kept in the summary. This diff retains `PropagateTaintFrom` attributes by calculating transitive taint-from information from addresses that are kept in the summary.

Reviewed By: skcho

Differential Revision: D37178425

fbshipit-source-id: d19bfe4ef2783d31dfc70258f8e5f147c865435d