[erl-frontend] Use existing equality function from Erlang
Summary:
We don't need a builtin decl for equality in Erlang, as there is actually an existing function in Erlang for it (similarly to other operators): `erlang:'=='/2`. Same for exactly equal. Previously we didn't treat equality and exactly equal differently, even in the translation. Now it's easy to treat them differently in the translation, so the issue is pushed down to Pulse.

Furthermore, if someone writes `erlang:'=='(1, 1)` (which is completely valid) instead of `1 == 1`, that will also work now.

Reviewed By: thizanne

Differential Revision: D37237336

fbshipit-source-id: dbd140342544642bf37a5519bb446f2f491417dd