[pulse][taint] Handle multiple sources/sanitizers/sinks
Summary:
Augment Pulse's taint analysis to handle multiple taint sources/sanitizers/sinks for a given address, and feed changes
through as necessary.

Aside from being a general addition to the taint analysis, this will be particularly useful when reporting data flows
from taint sources. In particular, a given tainted value can flow to multiple procedures (which are regarded as "sinks"
here), so we need to be able to represent all of these "sinks" reachable via an address in a single attribute
akin to `MustNotBeTainted`.

Reviewed By: skcho

Differential Revision: D37142491

fbshipit-source-id: 52ac87686f7c8ddb1fe4f1798939b34b992d93bd