Replace objects by records in Erlang comparison models
Summary:
Objects were working nicely for prototyping and for avoiding some record definitions, but now that
the code is settled and we have to define the method type anyway, it does not provide much
advantages. Yet it does have drawbacks, the main one being that it's less robust against typos in the
method names (defining an #integre method instead of #integer might not be detected, and due to the
inheritance mechanism the default imprecise method could be used).

We replace them by records which are virtually strictly better in all relevant regards.

Differential Revision: D37243650

fbshipit-source-id: 80c12959ac9bc1f448a3417a789a2495df65ac41