[misc] Don't run `opam user-setup` as part of `make devsetup`
Summary:
Arguably, people who want to hack on Infer already know (or motivated enough to learn to) how to configure their editor
of choice for OCaml.

At the same time `user-setup` can mess up an existing editor config, so running it in a non-interactive mode may not be
the best possible experience.

Reviewed By: skcho

Differential Revision: D37276654

fbshipit-source-id: 29ebff1dd5c8ac6189d9d065c58f1e745c2b2519