[pulse][taint] Improve checking of sanitisers along tainted flows
Summary:
Add timestamps to taint sources and sanitisers so that some attempt can be made at comparing their recency.
Previously when checking an incoming flow to a sink, only the *existence* of taint sources/santisers along the flow was checked.
Now, timestamps are used to try to determine an ordering to these too - i.e. if sanitisation happened *after* tainting.
In particular, if tainting is known to happen after sanitisation we should still report taint flows.

We use *traces* of timestamps - consider the following example, where a procedure both taints and sanitises:
  function foo(obj) {
    sanitize(obj);
    taint(obj);
  }

  function bar() {
    obj = new Object();
    foo(obj);
    sink(obj);
  }

When analysing `foo`, the `TaintSanitized` attribute will be older than the `Tainted` attribute.
However, during the interprocedural call to `foo` in `bar`, both inherit the timestamp of the call to `foo` itself.
We need to remember the timestamps from analysis of `foo` to order the attributes correctly - so we use a trace of
timestamps (i.e. a list with newest timestamps at the head), and compare the trace lexicographically to determine
whether tainting or sanitisation occurred first.

Reviewed By: da319

Differential Revision: D37380215

fbshipit-source-id: d8b628cf91fafc068ab554ae7a375830df0b5500