[pulse taint][c++] Import quandary/strings.cpp
Summary: Pulse to find all the taint flows that quandary does for `strings.cpp`. We needed to add special cases for some of the c++ functions/methods, similarly as quandary does. The only difference was the special case for the `swap` function. In Pulse, we do not propagate to receiver if the function is static.

Reviewed By: jvillard

Differential Revision: D37486977

fbshipit-source-id: 6f18659c419ed9e5edc456ac1128a17e282422ad